enable frontend-navigation, embedded-reactivity and protected-pointers experimental features per default
