+++ src/routing/frontend-routing.ts
@@ -1,171 +1,144 @@
-// deno-lint-ignore-file no-namespace
-
-import { Path } from "datex-core-legacy/utils/path.ts";
-import { resolveEntrypointRoute,  refetchRoute } from "./rendering.ts";
-import { Datex } from "datex-core-legacy";
-import { Entrypoint, html_content_or_generator } from "../html/entrypoints.ts";
-import { KEEP_CONTENT } from "../html/entrypoint-providers.tsx";
-import { displayError } from "../html/errors.tsx";
-import { domUtils } from "../app/dom-context.ts";
-import { PartialHydration } from "../hydration/partial-hydration.ts";
+import { Logger } from "datex-core-legacy/utils/logger.ts";
+import { Entrypoint } from "../html/entrypoints.ts";
+import { Path } from "../utils/path.ts";
 import { ContextBuilder } from "./context.ts";
+import { resolveEntrypointRoute } from "./rendering.ts";
+import { domUtils } from "../app/dom-context.ts";
 import { querySelector } from "../uix-dom/dom/shadow_dom_selector.ts";
-import { recreateGlobalStyleSheetLinks } from "../utils/css-style-compat.ts";
-import { recreatePersistentListeners } from "datex-core-legacy/utils/persistent-listeners.ts";
+import { Datex } from "datex-core-legacy/mod.ts";
+import { PartialHydration } from "../hydration/partial-hydration.ts";
+import { displayError } from "../html/errors.tsx";
+import { setCookie } from "../session/cookies.ts";
+import { app } from "../app/app.ts";
 
-/**
- * Generalized implementation for setting the route in the current tab URL
- * Used in combination with components
- * You should only use the Routing.update() method in most cases to update the current URL, and otherwise rely on the component specific routing implementation (resolveRoute, handleRoute, getInternalRoute)
- */
+// navigation api polyfill
+if (!(globalThis as any).navigation) {
+	await import("../lib/navigation-api-polyfill/navigation-api.js");
+}
 
-const logger = new Datex.Logger("UIX Routing");
+const logger = new Logger("frontend router");
 
-export namespace Routing {
+type MergeStrategy = 'override'|'insert'
 
-	let frontend_entrypoint: Entrypoint|undefined
-	let backend_entrypoint: Entrypoint|undefined
-	let current_entrypoint: Entrypoint|undefined
-	let current_content: any;
+const experimentalViewTransitionsEnabled = app.options?.experimental_features.includes("view-transitions")
 
-	// @deprecated
-	export const Prefix = {};
-	export function setPrefix(){}
+export class FrontendRouter {
 
-	export function getCurrentRouteFromURL() {
-		return Path.Route(window.location.href ?? import.meta.url);
-	}
+	#frontendEntrypoint: Entrypoint;
+	#backendEntrypoint: Entrypoint;
 
-	export function setCurrentRoute<S extends boolean>(url?:string|URL, silent?: S): S extends true ? boolean : Promise<boolean>
-	export function setCurrentRoute<S extends boolean>(parts?:string[], silent?: S): S extends true ? boolean : Promise<boolean>
-	export function setCurrentRoute(_route?:string|string[]|URL, silent = false) {
-		if (!globalThis.history) return false;
-		const route = Path.Route(_route);
-		if (Path.routesAreEqual(getCurrentRouteFromURL(), route)) return false; // no change, ignore
-
-		history.pushState(null, "", route.routename);
+	#currentContent: unknown;
+	#currentMergeStrategy?: MergeStrategy;
 	
-		if (!silent) return resolveCurrentRoute();
-		else return true;
+	setEntrypoints(frontend?: Entrypoint, backend?: Entrypoint, isHydrating = false, mergeStrategy: MergeStrategy = 'override', requestBackendContent = false) {
+		this.#frontendEntrypoint = frontend;
+		this.#backendEntrypoint = backend;
+
+		// no initial ssr content provided, request it from the backend
+		if (requestBackendContent) this.renderRouteContent();
+		// render current route frontend content, backend content is already provided from the HTTP response
+		else this.renderRouteFrontendContent(isHydrating, mergeStrategy);
 	}
 
-
-	export async function setEntrypoints(frontend?: Entrypoint, backend?: Entrypoint, isHydrating = false, mergeFrontend: 'override'|'insert'|undefined = 'insert') {
-		frontend_entrypoint = frontend;
-		backend_entrypoint = backend;
-		// entrypoints available - enable frontend routing
-		if (frontend_entrypoint || backend_entrypoint) {
-			enableFrontendRouting();
+	async navigateTo(route: string|URL) {
+		const routePath = new Path(route, window.location.origin);
+		if (new Path(location.href).toString() === routePath.toString()) {
+			logger.info("already on route " + route + ", skipping navigation");
+			return;
 		}
 
-		const backend_available = isHydrating || (backend_entrypoint ? await renderEntrypoint(backend_entrypoint) : false);
-		const frontend_available = (mergeFrontend || !backend_available) && frontend_entrypoint ? await renderEntrypoint(frontend_entrypoint, backend_available ? mergeFrontend : undefined) : false;
-
-		// no content for path found after initial loading
-		if (!frontend_available && !backend_available) {
-			if (!backend_entrypoint && !frontend_entrypoint) displayError("No entrypoint", `No backend or frontend entrypoint (default export) available.`)
-			else displayError("No content", `Route resolved to null on the ${backend_entrypoint?'backend':''}${(backend_entrypoint&&frontend_entrypoint?' and ': '')}${frontend_entrypoint?'frontend':''}`)
-		}
+		// update current tab url
+		history.pushState(null, "", routePath.routename);
+		// render content
+		await this.renderRouteContent(routePath);
 	}
 
-	export async function renderEntrypoint(entrypoint:Entrypoint, mergeType?: 'override'|'insert') {
-		const content = await getContentFromEntrypoint(entrypoint, undefined)
-		if (content != null && content !== KEEP_CONTENT) {
-			if (mergeType == "insert") {
+	async renderRouteContent(route: string|URL = location.href) {
+		const routePath = new Path(route, window.location.origin);
+		// render content
+		const {content: backendContent, status_code } = await this.getBackendContent(route);
 
-				let elements = [];
-				if (content instanceof DocumentFragment) elements = [...content.children]
-				else if (content instanceof Array) elements = content;
-				else if (content instanceof Response) {
-					// TODO:
-					await renderResponse(content)
-					logger.warn("Frontend entrypoint returned a Response object - this cannot be merged with the provided backend content");
-				}
-				else elements = [content];
-
-				for (const el of elements) {
-					let slot:HTMLElement|null = null;
-					if (el instanceof HTMLElement && el.hasAttribute("slot")) {
-						const name = el.getAttribute("slot")!
-						slot = querySelector(`frontend-slot[name="${domUtils.escapeHtml(name)}"]`) as HTMLElement;
-						if (!slot) logger.error(`Could not find a matching <frontend-slot name="${name}"/> for content provided from frontend entrypoint. Make sure your backend and frontend routes are not unintentionally colliding.`);
-					}
-					else {
-						slot = querySelector("frontend-slot") as HTMLElement;
-						if (!slot) logger.error("Could not find a matching <frontend-slot/> for content provided from frontend entrypoint. Make sure your backend and frontend routes are not unintentionally colliding.");
-					}
-	
-					if (slot) {
-						slot.innerHTML = "";
-						slot.append(el);
-					}
-				}				
+		// full reload if status code is 302
+		if (status_code === 302) {
+			if (typeof backendContent == "string") {
+				// set uix-cache-token cookie
+				setCookie("uix-cache-token", routePath.pathname + ":" + backendContent);
 			}
-			else await setContent(content, entrypoint)
+			globalThis.location.href = routePath.routename;
+			return;
 		}
-		return content != null
+
+		if (backendContent!=null) this.overrideContent(backendContent);
+
+		const isHydrating = backendContent != null;
+		const mergeStrategy = isHydrating ? 'insert' : 'override';
+		this.renderRouteFrontendContent(isHydrating, mergeStrategy, route);
 	}
 
+	async renderRouteFrontendContent(isHydrating: boolean, mergeStrategy: MergeStrategy = 'override', route: string|URL = location.href) {
+		// has backend content?
+		const hasBackendRenderedContent = isHydrating;
 
-	async function getContentFromEntrypoint(entrypoint: Entrypoint, route: Path.Route = getCurrentRouteFromURL(), probe_no_side_effects = false) {
-		// create new context with fake request
-		const url = new Path(route, window.location.origin);
-		const path = route.pathname;
-		const context = new ContextBuilder()
-			.setRequestData(new Request(url), path)
-			.build()
-		context.path = path;
-		context.request = new Request(url)
+		// no backend content, just render frontend content
+		if (!hasBackendRenderedContent) mergeStrategy = 'override';
 
-		const { content } = await resolveEntrypointRoute({entrypoint, context, route, probe_no_side_effects});
-		return content;
+		const {content} = await this.getContentFromEntrypoint(this.#frontendEntrypoint, route);
+
+		if (content == null) {
+			if (!hasBackendRenderedContent) displayError("UIX Rendering Error", "Route not found on backend or frontend");
+			this.#currentContent = content;
+			this.#currentMergeStrategy = mergeStrategy;
+			return;
+		}
+
+		if (content == this.#currentContent && mergeStrategy == this.#currentMergeStrategy) {
+			logger.info("content for " + route + " is the same, skipping");
+			return;
+		}
+
+		logger.success("navigating to " + route);
+
+		this.#currentContent = content;
+		this.#currentMergeStrategy = mergeStrategy;
+
+		if (mergeStrategy == 'override') this.overrideContent(content);
+		else if (mergeStrategy == 'insert') this.insertContent(content);
+		else throw new Error("Invalid merge strategy: " + mergeStrategy);
 	}
 
-	async function setContent(content: html_content_or_generator, entrypoint:Entrypoint) {
-		current_entrypoint = entrypoint;
+	overrideContent(content: unknown) {
+		if (content == null) return;
 
-		if (current_content !== content) {
-			current_content = content;
-			// TODO:
-			if (content == null) return;
-
-			// partial hydration, no need to set new dom nodes 
-			else if (content instanceof PartialHydration) {
-				return;
-			}
-
-			else if (content instanceof Array) {
-				document.body.innerHTML = "";
-				domUtils.append(document.body, content) // add to document
-			}
-			// handle response
-			else if (content instanceof Response) {
-				renderResponse(content)
-			}
-
-			// TODO: currently only displayed if type not correctly mapped (TypedValue fallback)
-			else if (!(content instanceof Datex.TypedValue)) { //if (content instanceof Element || content instanceof DocumentFragment) {
-				document.body.innerHTML = "";
-				// TODO: handle all content correctly (same behaviour as on backend)
-				domUtils.append(document.body, content) // add to document
-			}
-			else {
-				displayError("UIX Rendering Error", "Cannot render value of type " + Datex.Type.ofValue(content));
-			}
-			// else {
-			// 	logger.error("invalid content, cannot handle yet", content)
-			// 	return;
-			// }
+		// partial hydration, no need to set new dom nodes 
+		else if (content instanceof PartialHydration) {
+			return;
 		}
-	
-		await update(getCurrentRouteFromURL(), false)
+
+		else if (content instanceof Array) {
+			document.body.innerHTML = "";
+			domUtils.append(document.body, content) // add to document
+		}
+		// handle response
+		else if (content instanceof Response) {
+			this.renderResponse(content)
+		}
+
+		// TODO: currently only displayed if type not correctly mapped (TypedValue fallback)
+		else if (!(content instanceof Datex.TypedValue)) { //if (content instanceof Element || content instanceof DocumentFragment) {
+			document.body.innerHTML = "";
+			// TODO: handle all content correctly (same behaviour as on backend)
+			domUtils.append(document.body, content as any) // add to document
+		}
+		else {
+			displayError("UIX Rendering Error", "Cannot render value of type " + Datex.Type.ofValue(content));
+		}
 	}
 
 	/**
 	 * Render a Response on the client side
 	 * @param response 
 	 */
-	async function renderResponse(response: Response) {
+	async renderResponse(response: Response) {
 		if (response.body instanceof ReadableStream) {
 			// if (isContentType(response, "text/html")) {
 			// 	if (response.redirected) setCurrentRoute(response.url, true);
@@ -183,8 +156,7 @@ export namespace Routing {
 			// 	document.close();
 			// }
 			// else 
-			if (isContentType(response, "text/plain")) {
-				if (response.redirected) setCurrentRoute(response.url, true);
+			if (this.isContentType(response, "text/plain")) {
 				const content = await response.text()
 				document.body.innerHTML = '<pre style="all:initial;word-wrap: break-word; white-space: pre-wrap;">'+domUtils.escapeHtml(content)+'</pre>'
 			}
@@ -193,7 +165,8 @@ export namespace Routing {
 			}
 		}
 		else if (response.status === 302) {
-			window.location.href = response.headers.get("location")!;
+			this.navigateTo(response.headers.get("location")!);
+			// window.location.href = response.headers.get("location")!;
 		}
 		else if (response.body) {
 			console.warn("cannot handle response body on frontend (TODO)", response)
@@ -203,105 +176,85 @@ export namespace Routing {
 		}
 	}
 
-	function isContentType(response: Response, mimeType: `${string}/${string}`) {
+	isContentType(response: Response, mimeType: `${string}/${string}`) {
 		const actualMimeType = response.headers.get("content-type") 
 		return actualMimeType === mimeType || actualMimeType?.startsWith(mimeType + ";")
 	}
 
-	async function resolveCurrentRouteFrontend(){
-		// try frontend entrypoint
-		if (frontend_entrypoint) {
-			const content = await getContentFromEntrypoint(frontend_entrypoint)
-			if (content !== null) {
-				const entrypoint = frontend_entrypoint;
-				return {content, entrypoint}
+	insertContent(content: unknown) {
+		// get elements list from content
+		const elements = this.getElementsListFromContent(content);
+
+		// no elements could be extracted
+		if (!elements) {
+			if (content instanceof Response) {
+				logger.warn("Content of type 'Response' cannot be merged with existing content, falling back to override");
+				this.overrideContent(content)
+				return;
 			}
 			else {
-				displayError("UIX Rendering Error", "Route not found on backend or frontend");
-				return {}
+				logger.warn("content is not valid");
+				return;
 			}
 		}
-		else {
-			displayError("UIX Rendering Error", "Route not found on backend or frontend");
-			return {};
-		}
-	}
+		
+		// insert into existing DOM
+		for (const el of elements) {
+			let slot:HTMLElement|null = null;
+			if (el instanceof HTMLElement && el.hasAttribute("slot")) {
+				const name = el.getAttribute("slot")!
+				slot = querySelector(`frontend-slot[name="${domUtils.escapeHtml(name)}"]`) as HTMLElement;
+				if (!slot) logger.error(`Could not find a matching <frontend-slot name="${name}"/> for content provided from the frontend entrypoint. Make sure your backend and frontend routes are not unintentionally colliding.`);
+			}
+			else {
+				slot = querySelector("frontend-slot") as HTMLElement;
+				if (!slot) return this.overrideContent(content);
+			}
 
-	async function resolveCurrentRoute(allowReload=true){
-		// fix: disable backend response fetching completely
-		if (!allowReload) console.warn("ignoring allowReload=false, reloading page");
-		window.location.reload();
-	}
-
-	function getInferredDOMEntrypoint(){
-		return document.body.children[0] instanceof Element ? document.body.children[0] : null
-	}
-
-	const INITIAL_LOAD = Symbol("INITIAL_LOAD")
-
-	/**
-	 * updates the current URL with the current route requested from the get_handler
-	 */
-	export async function update(route_should_equal?:Path.route_representation, load_current_new = false){
-
-		const current = getCurrentRouteFromURL();
-		const route = route_should_equal ?? current;
-
-		// first load current route
-		if (load_current_new) await resolveCurrentRoute();
-
-		let changed = !!route_should_equal;
-
-		const usingInferredEntrypoint = !current_entrypoint; // reconstructing entrypoint from DOM. Probable reason: content was server side rendered
-		const entrypoint = current_entrypoint// ?? getInferredDOMEntrypoint();
-
-		if (entrypoint) {
-			// entrypoint was inferred but inferred entrypoint was not yet initially routed
-			if (usingInferredEntrypoint) {
-				const loadedInitial = entrypoint[INITIAL_LOAD];
-				entrypoint[INITIAL_LOAD] = true;
-				if (!loadedInitial) await resolveCurrentRoute();
-			} 
-
-			// TODO: use route refetching?
-			// const refetched_route = await refetchRoute(route, entrypoint);// Path.Route(await (<RouteManager>current_content).getInternalRoute());
-			// // check if accepted route matches new calculated current_route
-			// if (route_should_equal && !Path.routesAreEqual(route_should_equal, refetched_route)) {
-			// 	logger.warn `new route should be "${Path.Route(route_should_equal).routename}", but was changed to "${refetched_route.routename}". Make sure getInternalRoute() and onRoute() are consistent in all components.`;
-			// 	// stop ongoing loading animation
-			// 	window.stop()
-			// 	// something is wrong, content was server side rendered, routes might not be resolved correctly, better reload to get server routing
-			// 	if (usingInferredEntrypoint) window.location.href = Path.Route(route_should_equal).routename; 
-			// }
-
-			// // must be updated to new
-			// if (!Path.routesAreEqual(current, refetched_route)) {
-			// 	changed = setCurrentRoute(refetched_route, true); // update silently
-			// }
+			if (slot) {
+				slot.innerHTML = "";
+				domUtils.append(slot, el);
+			}
 		}
 
-		// if (changed) logger.success `new route: ${getCurrentRouteFromURL().routename??"/"}`;
 
 	}
 
-	let frontendRoutingEnabled = false;
+	private getElementsListFromContent(content: unknown) {
+		let elements = [];
+		if (content instanceof DocumentFragment) elements = [...content.children as unknown as HTMLElement[]]
+		else if (content instanceof Array) elements = content;
+		else if (content instanceof Response) return null;
+		else elements = [content];
+		return elements;
+	}
 
-	function enableFrontendRouting() {
 
-		// TODO: fix
-		return;
+	async getContentFromEntrypoint(entrypoint: Entrypoint, route: URL|string, probe_no_side_effects = false) {
 
-		if (frontendRoutingEnabled) return;
-		frontendRoutingEnabled = true;
-		logger.debug("frontend routing enabled");
+		route = route instanceof Path ? route : new Path(route, window.location.origin);
 
-		// listen for history changes
+		// create new context with fake request
+		const url = new Path(route, window.location.origin);
+		const path = route.pathname;
+		const context = new ContextBuilder()
+			.setRequestData(new Request(url), path)
+			.build()
+		context.path = path;
+		context.request = new Request(url)
 
-		// TODO: reenable, currently only works in chrome and has some issues with hydration
-		// @ts-ignore
-		if (false && globalThis.navigation) {
-			// @ts-ignore
-			globalThis.navigation?.addEventListener("navigate", (e:any)=>{
+		const { content, status_code } = await resolveEntrypointRoute({entrypoint, context, route, probe_no_side_effects});
+		return { content, status_code };
+	}
+
+	getBackendContent(route: URL|string) {
+		return this.getContentFromEntrypoint(this.#backendEntrypoint, route);
+	}
+
+	enableNavigationInterception() {
+		const _globalThis = globalThis as any;
+		if (_globalThis.navigation) {
+			_globalThis.navigation?.addEventListener("navigate", (e:any)=>{
 				
 				if (!e.userInitiated || !e.canIntercept || e.downloadRequest || e.formData) return;
 				
@@ -313,22 +266,24 @@ export namespace Routing {
 
 				// TODO: this intercept should be cancelled/not executed when the route is loaded from the server (determined in handleCurrentURLRoute)
 				e.intercept({
-					async handler() {
-						await resolveCurrentRoute();
+					handler: () => {
+						// render content
+						// With view transitions
+						if (experimentalViewTransitionsEnabled && document.startViewTransition) document.startViewTransition(() => this.renderRouteContent(url));
+						// Without view transitions
+						else return this.renderRouteContent(url);
 					},
 					focusReset: 'manual',
 					scroll: 'manual'
 				})
-				e.s
-			})
-		}
-
-		// fallback if "navigate" event not supported - only works for # paths, otherwise, page is reloaded
-		else {
-			globalThis.addEventListener('popstate', (e) => {
-				resolveCurrentRoute();
 			})
 		}
 	}
 
 }
+
+export const frontendRouter = new FrontendRouter();
+
+frontendRouter.enableNavigationInterception();
+
+globalThis.frontendRouter = frontendRouter;
